moved stories classes into stories.css.ts
