+++ packages/libs/react-ui/src/components/Grid/stories.css.ts
@@ -1,0 +1,19 @@
+import { sprinkles } from '../../styles';
+
+import { style } from '@vanilla-extract/css';
+
+export const ContentClass = style([
+  sprinkles({
+    backgroundColor: 'primarySurface',
+    borderRadius: 'sm',
+    padding: 2,
+    color: 'neutral6',
+    display: 'flex',
+  }),
+  {
+    width: '100%',
+    height: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+]);
